+++ BALSAMIC/commands/config/case.py
@@ -33,6 +33,8 @@ from BALSAMIC.commands.options import (
     OPTION_PANEL_BED,
     OPTION_PON_CNN,
     OPTION_QUALITY_TRIM,
+    OPTION_SENTIEON_INSTALL_DIR,
+    OPTION_SENTIEON_LICENSE,
     OPTION_SWEGEN_SNV,
     OPTION_SWEGEN_SV,
     OPTION_TUMOR_SAMPLE_NAME,
@@ -84,6 +86,8 @@ LOG = logging.getLogger(__name__)
 @OPTION_PANEL_BED
 @OPTION_PON_CNN
 @OPTION_QUALITY_TRIM
+@OPTION_SENTIEON_INSTALL_DIR
+@OPTION_SENTIEON_LICENSE
 @OPTION_SWEGEN_SNV
 @OPTION_SWEGEN_SV
 @OPTION_TUMOR_SAMPLE_NAME
@@ -117,6 +121,8 @@ def case_config(
     panel_bed: Path,
     pon_cnn: Path,
     quality_trim: bool,
+    sentieon_install_dir: Path,
+    sentieon_license: str,
     swegen_snv: Path,
     swegen_sv: Path,
     tumor_sample_name: str,
@@ -188,6 +194,8 @@ def case_config(
         directory.mkdir(exist_ok=True)
 
     config_collection_dict = ConfigModel(
+        sentieon_install_dir=sentieon_install_dir if sentieon_install_dir else None,
+        sentieon_license=sentieon_license if sentieon_license else None,
         QC={
             "quality_trim": quality_trim,
             "adapter_trim": adapter_trim,